# Add Time#to_json method to improve performance