When it pass an object containing Time objct as an argument of `JSON.generate`,
the following processes are performed.

1. it will invoke `Time#to_json`
   at https://github.com/flori/json/blob/3d1c6a9246251e382c80dd8d85844c0bbf182da1/ext/json/ext/generator/generator.c#L1019
2. `mObject_to_json` function will be called by invoking `Time#to_json`.
3. it will invoke `Time#to_s`
   at https://github.com/flori/json/blob/3d1c6a9246251e382c80dd8d85844c0bbf182da1/ext/json/ext/generator/generator.c#L576

Seems that it reducde the performance by two Ruby method calls using `rb_funcall`.

This patch adds `generate_json_time` and makes a direct call to `Time#to_s` to reduce `rb_funcall` calling.
The performance is improved by about 20 %.

## Environment
- MacBook Pro (16-inch, 2019)
- macOS 10.15.5
- Intel Core i9 2.4 GHz

### Before
```
Warming up --------------------------------------
       JSON.generate    16.650k i/100ms
        Time.to_json    20.176k i/100ms
Calculating -------------------------------------
       JSON.generate    167.636k (± 0.9%) i/s -    849.150k in   5.065832s
        Time.to_json    200.217k (± 1.2%) i/s -      1.009M in   5.039224s
```

### After
```
Warming up --------------------------------------
       JSON.generate    20.297k i/100ms
        Time.to_json    20.329k i/100ms
Calculating -------------------------------------
       JSON.generate    205.745k (± 1.1%) i/s -      1.035M in   5.031798s
        Time.to_json    203.763k (± 0.7%) i/s -      1.037M in   5.088436s
```

### Test code
```ruby
require 'json'
require 'benchmark/ips'

time = Time.now

Benchmark.ips do |x|
  x.report('JSON.generate') { JSON.generate(time) }
  x.report('Time.to_json') { time.to_json }
end
```