@marcandre 2020-06-28
Cool PR. I find that gain surprising. I'm not against optimizing for `Time`, but it would be even nicer to optimize for all objects without a specialized `to_json`...

So one thing I notice in `generate_json` is the call to `rb_respond_to(obj, i_to_json)`. Since this will be true for the vast majority of cases, couldn't we instead simply call `to_json` on the `obj` and only in case we get a `NoMethodError` have the fallback that follows?

In Ruby:
```ruby
# generate_json
case
when obj.is_a?(Hash)
  # ...
when obj.respond_to?(:to_json)
  obj.to_json
else
  obj.to_s
end
# change to:
case
when obj.is_a?(Hash)
  # ...
else
  begin
    obj.to_json
  rescue NoMethodError
    raise if obj.repond_to?(:to_json) # must be an error within a custom `#to_json`
    obj.to_s
  end
end
```

I'm curious to know much of that 20% gain is due to this avoidable call to `respond_to?`?