@marcandre 2020-06-28
And now that I'm thinking about it, it's a bit odd that there's not a `rb_call_if_respond_to` or similar, to avoid this pattern of looking up for a method call twice.