@Watson1978 2020-06-29
```
  begin
    obj.to_json
  rescue NoMethodError
    raise if obj.repond_to?(:to_json) # must be an error within a custom `#to_json`
    obj.to_s
  end
```

In the case of calling #to_s, I'm guessing that performance will be worse than currently.
(I think it is big overhead to catch the exceptions.)

So, I think it would be better to verify that in another PR.