@marcandre 2020-06-29
> In the case of calling #to_s,

No, in the case of `generate_json` which will not call `to_s` unless the object inherits only from BasicObject or has `undef to_json`.

> (I think it is big overhead to catch the exceptions.)

I'm sorry if I wasn't clear, but we wouldn't be catching exceptions (except in the very unusual cases of an object that inherits only from BasicObject or has `undef to_json`).

Or am I missing something?